fix use of uint

+++ pkg/database/database_test.go
@@ -8,8 +8,7 @@ import (
 
 func TestDatabase(t *testing.T) {
 	type Person struct {
-		// TODO uint64 obvsly
-		ID   uint   `attr:"sys/db/id"`
+		ID   uint64 `attr:"sys/db/id"`
 		Name string `attr:"person/name"`
 	}
 
@@ -28,7 +27,7 @@ func TestDatabase(t *testing.T) {
 		snapshot := BuildTypedSnapshot(res.Snap, &Person{})
 		person, ok := snapshot.Find(res.IDs[0])
 		assert.True(t, ok)
-		assert.Equal(t, Person{ID: uint(res.IDs[0]), Name: "Donald"}, *person)
+		assert.Equal(t, Person{ID: res.IDs[0], Name: "Donald"}, *person)
 	})
 
 	t.Run("find different struct with overlapping field", func(t *testing.T) {
@@ -46,15 +45,15 @@ func TestDatabase(t *testing.T) {
 		id := res.IDs[0]
 		snapshot1 := BuildTypedSnapshot(res.Snap, &Person{})
 		res = db.Write(Request{
-			Assertions: []any{Person{ID: uint(id), Name: "Donato"}},
+			Assertions: []any{Person{ID: id, Name: "Donato"}},
 		})
 		assert.NoError(t, res.Error)
 		snapshot2 := BuildTypedSnapshot(res.Snap, &Person{})
 		person, ok := snapshot2.Find(id)
 		assert.True(t, ok)
-		assert.Equal(t, Person{ID: uint(id), Name: "Donato"}, *person)
+		assert.Equal(t, Person{ID: id, Name: "Donato"}, *person)
 		person, ok = snapshot1.Find(id)
 		assert.True(t, ok)
-		assert.Equal(t, Person{ID: uint(id), Name: "Donald"}, *person)
+		assert.Equal(t, Person{ID: id, Name: "Donald"}, *person)
 	})
 }
